move tokenManager to api module
